# refactor: optimize the metadata.Equal for small number of finalizers