It might be pretty common to have 1-2 finalizers, so we can do a special case for it.

```
name                old time/op    new time/op    delta
MetadataEqual/0-32    12.1ns ± 2%    11.8ns ± 2%      ~     (p=0.057 n=4+4)
MetadataEqual/1-32     206ns ± 8%      14ns ± 0%   -93.11%  (p=0.029 n=4+4)
MetadataEqual/2-32     268ns ± 7%      21ns ± 2%   -92.13%  (p=0.029 n=4+4)
MetadataEqual/3-32     328ns ± 4%     227ns ± 4%   -30.86%  (p=0.029 n=4+4)

name                old alloc/op   new alloc/op   delta
MetadataEqual/0-32     0.00B          0.00B           ~     (all equal)
MetadataEqual/1-32     80.0B ± 0%      0.0B       -100.00%  (p=0.029 n=4+4)
MetadataEqual/2-32      112B ± 0%        0B       -100.00%  (p=0.029 n=4+4)
MetadataEqual/3-32      144B ± 0%       96B ± 0%   -33.33%  (p=0.029 n=4+4)

name                old allocs/op  new allocs/op  delta
MetadataEqual/0-32      0.00           0.00           ~     (all equal)
MetadataEqual/1-32      4.00 ± 0%      0.00       -100.00%  (p=0.029 n=4+4)
MetadataEqual/2-32      4.00 ± 0%      0.00       -100.00%  (p=0.029 n=4+4)
MetadataEqual/3-32      4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.029 n=4+4)
```

Fixes #306