@DmitriyMV 2023-09-22
Small idea - we can actually speed up cases up to 16 (after 16 elements ns/op decrease drops below 50%).

```go
func equalUpTo16(a, b []Finalizer) bool {
	if len(a) != len(b) {
		return false
	}

	if len(a) > 16 {
		panic("len(a) is greater than 16")
	}

	var elementCompared uint16

	for i := range a {
		idx := -1

		for j := range b {
			if elementCompared& (1 << j) == 0 && a[i] == b[j] {
				idx = j

				break
			}
		}

		if idx == -1 {
			return false
		}

		elementCompared |= 1 << idx
	}

	return true
}
```

And use it like 
```go
if len(fins) <= 16 {
	equalUpTo16(fins, otherFins)
}
```