@smira 2023-09-22
> Small idea - we can actually speed up cases up to 16 (after 16 elements ns/op decrease drops below 50%).

I feel this is a bit of overkill? I mean readability vs. optimization :)