@DmitriyMV 2023-09-22
> I feel this is a bit of overkill? I mean readability vs. optimization :)

I mean, probably? Depends on how often we see resources with 3-16 finalizers. Interestingly enough `equalUpTo16` is almost as fast for cases of 2 and 3 elements as current direct comparison.

What is more interesting `equalUpTo64` is 20% slower than current implementation with Clone,Sort,Equal.