@smira 2023-09-22
Updated with @DmitriyMV suggestions, I feel 1-2 finalizers version is good enough, and also updated >= 3 finalizers version is better than before.